% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{basal_group}
\alias{basal_group}
\title{Basal Group}
\usage{
basal_group(
  value,
  lookup,
  hierarchical_variable,
  nesting_variable,
  encountered_values = character()
)
}
\arguments{
\item{value}{Value (e.g., \code{hepatitis-A}) of the \code{hierarchical_variable}
(e.g., \code{disease}) for which to determine basal value (e.g., \code{hepatitis}).}

\item{lookup}{Table with two character-valued columns with names given by
\code{hierarchical_variable} (e.g., \code{disease}) and \code{nesting_variable}
(e.g., \code{nesting_disease}).}

\item{hierarchical_variable}{Name of the hierarchical variable in \code{lookup}}

\item{nesting_variable}{Name of the nesting variable in \code{lookup}}

\item{encountered_values}{Character vector of values already found.
Typically this left at the default value of an empty character vector.}
}
\value{
The root value that input value maps to in \code{lookup}.
}
\description{
Basal Group
}
