% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{BasalGroupAdder}
\alias{BasalGroupAdder}
\title{Basal Group Adder}
\usage{
BasalGroupAdder(lookup)
}
\arguments{
\item{lookup}{A lookup table with \code{disease} and \code{nesting_disease}
columns that describe a global disease hierarchy that will be applied
to find the basal disease of each \code{disease} in data.}

\item{data}{A tidy data set with a \code{disease} column.}
}
\value{
tidy dataset with basal disease
}
\description{
Basal Group Adder
}
