% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_dependencies.R
\name{browse_pipeline_dependencies}
\alias{browse_pipeline_dependencies}
\title{Browse Pipeline Dependencies}
\usage{
browse_pipeline_dependencies(
  dataset_ids,
  dependency_types = c("IsCompiledBy", "IsDerivedFrom", "References"),
  metadata = iidda.api::ops$metadata(dataset_ids = dataset_ids)
)
}
\arguments{
\item{dataset_ids}{Character vector of dataset identifiers.}

\item{dependency_types}{Vector of types of dependencies to browse. Possible
values include \code{"IsCompiledBy"}, \code{"IsDerivedFrom"}, and
\code{"References"}.}

\item{metadata}{Optional list giving dataset metadata. The default uses
the IIDDA API, which requires the internet.}
}
\description{
Open a browser at the locations of the dependencies associated with a set
of datasets.
}
