% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{ca_iso_3166_2}
\alias{ca_iso_3166_2}
\title{Order Canadian Provinces Geographically}
\usage{
ca_iso_3166_2(data)
}
\arguments{
\item{data}{Dataset containing an \code{iso_3166_2} field with Canadian province
and territory codes.}
}
\description{
Order Canadian Provinces Geographically
}
\concept{normalization}
