% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{check_date}
\alias{check_date}
\title{Test if x is a Date, coerce if not}
\usage{
check_date(x)
}
\arguments{
\item{x}{vector of putative dates}
}
\value{
vector with class Date, or error
}
\description{
Test if x is a Date, coerce if not
}
\examples{
d1 <- check_date("1920-01-01")
d1
class(d1)
# returns an error if x can't be coerced to Date easily
# check_date("may 29th")
}
