% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{ComputeMovingAverage}
\alias{ComputeMovingAverage}
\title{Compute Moving Average of Time Series}
\usage{
ComputeMovingAverage(ma_window_length = 52)
}
\arguments{
\item{ma_window_length}{length of moving average window, this will depend on the time scale in the data.
Defaults to 52, so that weekly data is averaged over years.}
}
\value{
a function to remove to compute the moving average of a time series variable
\subsection{Returned Function}{
\itemize{
\item Arguments
\itemize{
\item \code{data} data frame containing time series data
\item \code{series_variable} column name of series variable in \code{data}, default is "deaths"
\item \code{time_variable} column name of time variable in \code{data}, default is "period_end_date"
}
\item Return - all fields in \code{data} with the \code{series_variable} data replaced with the moving
average.
}
}
}
\description{
Compute Moving Average of Time Series
}
\concept{data_prep_constructors}
