% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{CountAggregator}
\alias{CountAggregator}
\title{Count Aggregator}
\usage{
CountAggregator()
}
\value{
A function like \code{\link{count_aggregator_default}} that
aggregates count variables.
}
\description{
Create a function that aggregates count variables.
}
\concept{data_prep_constructors}
