% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{create_bin_desc}
\alias{create_bin_desc}
\title{Create age bin descriptions}
\usage{
create_bin_desc(age_df)
}
\arguments{
\item{age_df}{data frame of data with age_group column}
}
\value{
data frame of data with bin_desc column
}
\description{
Create age bin descriptions for joining age_group lookup table
}
\concept{harmonization}
