% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{CustomDataPrep}
\alias{CustomDataPrep}
\title{Custom Data Prep Function}
\usage{
CustomDataPrep(
  data_prep_function,
  ign_variables = character(0L),
  opt_variables = character(0L),
  new_variables = character(0L)
)
}
\arguments{
\item{data_prep_function}{A standard \code{R} function that takes a data frame,
called \code{data} as its first argument, and returns another data frame.
Optionally, subsequent arguments can be added that give the names
of types of variables (e.g., \code{period_end_variable = "period_end_date"}).
Consider using \code{NULL} as the default of these variable name arguments,
which will lead to good guesses about variable names when \code{data} are
obtained from \code{iidda.api}.}
}
\value{
description A data prep function that maintains
\code{iidda} attributes and guesses at names of types of variables that the
\code{data_prep_function} assumes.
}
\description{
Convert a \code{data_prep_function}, into
a function that can be used in an \code{iidda} data prep pipeline.
}
