% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R, R/std.R
\name{handle_missing_values_default}
\alias{handle_missing_values_default}
\alias{handle_zero_values_default}
\alias{trim_series_default}
\alias{series_harmonizer_default}
\alias{deheaper_default}
\alias{compute_moving_average_default}
\alias{period_aggregator_default}
\alias{count_aggregator_default}
\alias{period_describer_default}
\alias{data_prep_default}
\title{Data Prep Default Functions}
\usage{
handle_missing_values_default(data, series_variable = NULL)

handle_zero_values_default(data, series_variable = NULL)

trim_series_default(data, series_variable = NULL, time_variable = NULL)

series_harmonizer_default(data, series_variable = NULL, time_variable = NULL)

deheaper_default(data, series_variable = NULL, time_variable = NULL)

compute_moving_average_default(
  data,
  series_variable = NULL,
  time_variable = NULL
)

period_aggregator_default(
  data,
  time_variable = NULL,
  period_width_variable = NULL,
  count_variable = NULL,
  norm_variable = NULL
)

count_aggregator_default(
  data,
  total_count_variable = NULL,
  count_variable = NULL,
  grouping_variable = NULL
)

period_describer_default(
  data,
  period_start_variable = NULL,
  period_end_variable = NULL,
  period_mid_time_variable = NULL,
  period_mid_date_variable = NULL,
  period_days_variable = NULL
)
}
\arguments{
\item{data}{Data frame that likely comes from IIDDA.}

\item{series_variable}{Name of variable that can be used on the y-axis of a time-series.}

\item{time_variable}{Name of variable that characterizes the temporal location of the time period.}

\item{period_width_variable}{Name of variable that characterizes the width of the time period.}

\item{count_variable}{Name of variable containing a count (e.g., cases, births, deaths, population).}

\item{norm_variable}{Name of variable that can be used to normalize another variable (e.g., population normalized cases).}

\item{total_count_variable}{Name of variable containing a marginal total of a set of counts.}

\item{grouping_variable}{Name of variable containing}

\item{period_start_variable}{Name of variable containing}

\item{period_end_variable}{Name of variable containing}

\item{period_mid_time_variable}{Name of variable containing}

\item{period_mid_date_variable}{Name of variable containing}

\item{period_days_variable}{Name of variable containing}

\item{cases_variable}{Name of variable containing unstandardized reported incidence.}

\item{population_variable}{Name of variable containing population numbers.}

\item{birth_variable}{Name of variable containing numbers of births.}

\item{death_variable}{Name of variable containing numbers of deaths.}

\item{median_cases_variable}{Name of variable containing the median of a set of count variables.}

\item{period_mid_variable}{Name of variable containing}

\item{date_variable}{Name of variable containing}

\item{integer_time_variable}{Name of variable containing}

\item{numeric_time_variable}{Name of variable containing}

\item{time_scale_variable}{Name of variable containing}

\item{time_group_variable}{Name of variable containing}

\item{time_grouping_variable}{Name of variable containing}

\item{disease_variable}{Name of variable containing}

\item{hierarchical_variable}{Name of variable containing}

\item{nesting_variable}{Name of variable containing}

\item{basal_variable}{Name of variable containing}

\item{title_variable}{Name of variable containing}

\item{among_panel_variable}{Name of variable containing}

\item{within_panel_variable}{Name of variable containing}

\item{colour_variable}{Name of variable containing}

\item{category_variable}{Name of variable containing}

\item{categorical_variable}{Name of variable containing}
}
\description{
Data Prep Default Functions
}
