% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{DataDictionaryConverter}
\alias{DataDictionaryConverter}
\title{Data Dictionary Converter}
\usage{
DataDictionaryConverter(data_dictionary = iidda_data_dictionary())
}
\description{
Data Dictionary Converter
}
