% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{Deheaper}
\alias{Deheaper}
\title{De-heaping time series}
\usage{
Deheaper(
  prefix = "deheaped_",
  first_date = "1830-01-01",
  last_date = "1841-12-31",
  week_start = 45,
  week_end = 5,
  deheaping_scale = 2.7
)
}
\arguments{
\item{first_date}{string containing earliest date to look for heaping errors}

\item{last_date}{string containing last date to look for heaping errors}

\item{week_start}{numeric value of the first week number to start looking for heaping errors}

\item{week_end}{numeric value of the last week number to look for heaping errors}
}
\value{
function to fix heaping errors
\subsection{Returned Function}{
\itemize{
\item Arguments
\itemize{
\item \code{data} data frame containing time series data
}
\item Return - all fields in \code{data} with an additional field called "deheaped_" concatenated with \code{series_variable}.
If no heaping errors are found, this additional field is identical to the field `series_variable
}
}
}
\description{
Fixes heaping errors in time series. The structure of this function was taken from
the function \code{find_heap_and_deheap} created by Kevin Zhao
(https://github.com/davidearn/KevinZhao/blob/main/Report/make_SF_RData.R). This needs
to be better documented.
}
\concept{data_prep_constructors}
