% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{factor_time_scale}
\alias{factor_time_scale}
\title{Factor Time Scale}
\usage{
factor_time_scale(data)
}
\arguments{
\item{data}{A tidy data set with a \code{time_scale} column.}
}
\value{
A data set with a factored time_scale column.
}
\description{
Factor Time Scale
}
\concept{normalization}
