% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{find_unaccounted_cases}
\alias{find_unaccounted_cases}
\title{Find Unaccounted Cases}
\usage{
find_unaccounted_cases(data)
}
\arguments{
\item{data}{A tidy data set with a `basal_disease` column.}
}
\value{
A data set containing records that are the difference between a
reported total for a basal_disease and the sum of their leaf diseases.
}
\description{
Make new records for instances when the sum of leaf diseases is less than
the reported total for their basal disease. The difference between these
counts gets disease name 'basal_disease'_unaccounted'.
}
\concept{normalization}
