% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{generate_empty_df}
\alias{generate_empty_df}
\title{Create empty table with column names}
\usage{
generate_empty_df(dir_path, lookup_table, csv_name)
}
\arguments{
\item{dir_path}{string indicating path to directory}

\item{lookup_table}{data frame with column names to include in table}

\item{csv_name}{string indicating name of the created .csv file}
}
\value{
empty lookup table with columns from \code{lookup_table} in the directory if successfully generated
}
\description{
Creates an empty table in a specified directory
using columns names from another data frame
}
