% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{get_iidda_attr}
\alias{get_iidda_attr}
\title{Get IIDDA Attribute}
\usage{
get_iidda_attr(data, which)
}
\arguments{
\item{data}{Data frame that contains a list attribute called \code{"iidda"}.}

\item{which}{Name of the element in the \code{"iidda"} list to extract.}
}
\value{
Value of the element given by \code{which} in the \code{"iidda"} attribute.
}
\description{
Get IIDDA Attribute
}
