% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{get_implied_zeros}
\alias{get_implied_zeros}
\title{Get Implied Zeros}
\usage{
get_implied_zeros(data)
}
\arguments{
\item{data}{A tidy data set with the following minimal set of columns:
\code{disease}, \code{nesting_disease}, \code{year}, \code{original_dataset_id}, \code{iso_3166_2},
\code{basal_disease}, \code{time_scale}, \code{period_start_date}, \code{period_end_date},
\code{period_mid_date}, \code{days_this_period}, \code{dataset_id}}
}
\value{
A tidy data set with inferred 0s.
}
\description{
Add zeros to data set that are implied by a '0' reported at a coarser timescale.
}
\concept{normalization}
