% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{get_unit_labels}
\alias{get_unit_labels}
\title{Get time unit labels}
\usage{
get_unit_labels(unit)
}
\arguments{
\item{unit}{time unit, one of iidda.analysis:::time_units}
}
\value{
label of associated time unit
}
\description{
Get label of associated time unit
}
