% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{grid_dates}
\alias{grid_dates}
\title{Create a grid of dates starting at the first day in grid unit}
\usage{
grid_dates(
  start_date = "1920-01-01",
  end_date = "2020-01-01",
  by = "1 week",
  unit = "week",
  lookback = TRUE,
  week_start = 7
)
}
\arguments{
\item{start_date}{starting date}

\item{end_date}{end date}

\item{by}{increment of the sequence.  Optional.  See \sQuote{Details}.}

\item{unit}{a string, \code{Period} object or a date-time object. When a singleton string,
it specifies a time unit or a multiple of a unit to be rounded to. Valid base units
are \code{second}, \code{minute}, \code{hour}, \code{day}, \code{week}, \code{month}, \code{bimonth}, \code{quarter},
\code{season}, \code{halfyear} and \code{year}. Arbitrary unique English abbreviations as in the
\code{\link[lubridate:period]{period()}} constructor are allowed. Rounding to multiples of units (except weeks)
is supported. When \code{unit} is a \code{Period} object, units of the period objects are
used. This is equivalent to converting the period object to its string
representation and passing as \code{unit} argument.

When \code{unit} is a date-time object rounding is done to the nearest of the
elements in \code{unit}. If range of \code{unit} vector does not cover the range of
\code{x} \code{ceiling_date()} and \code{floor_date()} round to the \code{max(x)} and \code{min(x)}
for elements that fall outside of \code{range(unit)}.}

\item{lookback}{Logical, should the first value start before `start_date`}

\item{week_start}{week start day (Default is 7, Sunday. Set \code{lubridate.week.start} to
override). Full or abbreviated names of the days of the week can be in English or
as provided by the current locale.}
}
\value{
vector of Dates at the first of each week, month, year
}
\description{
Wrapper of `seq.Date()` and `lubridate::floor_date`
}
\examples{
grid_dates(start_date = "2023-04-01"
, end_date = "2023-05-16")

grid_dates(start_date = "2023-04-01"
, end_date = "2023-05-16"
, lookback = FALSE)


grid_dates(start_date = "2020-04-01"
, end_date = "2023-05-16"
, by = "2 months"
, unit = "month")
grid_dates(start_date = "2020-04-01"
, end_date = "2023-05-16"
, by = "2 months")
}
