% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{HandleMissingValues}
\alias{HandleMissingValues}
\title{Handle Missing Values in Series Variable}
\usage{
HandleMissingValues(na_remove = FALSE, na_replace = NULL)
}
\arguments{
\item{na_remove}{boolean value, if \code{TRUE} remove \code{NA}s in series variable}

\item{na_replace}{numeric value to replace \code{NA}s in series variable, if NULL
no replacement is performed}
}
\value{
A function like \code{\link{handle_missing_values_default}} that
removes or replaces missing values.
}
\description{
Construct a function that takes a data frame and returns another data
frame with \code{NA} values either removed or replaced.
}
\concept{data_prep_constructors}
