% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_attach_bar}
\alias{iidda_attach_bar}
\title{Attach Bar Plot to Dataset}
\usage{
iidda_attach_bar(
  data,
  initial_ggplot_object = ggplot(),
  series_variable = NULL,
  time_unit = NULL,
  aggregated = NULL
)
}
\arguments{
\item{data}{Data frame, probably containing an IIDDA dataset.}

\item{initial_ggplot_object}{Plot object that will be used to add a
bar geom.}
}
\description{
Attach Bar Plot to Dataset
}
