% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_get_metadata}
\alias{iidda_get_metadata}
\title{Get LBoM metadata}
\usage{
iidda_get_metadata(
  data,
  time_variable = "period_start_date",
  descriptor_variable = "cause"
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{time_variable}{column name of time variable in `data`, default is "period_start_date"}

\item{descriptor_variable}{column name of the descriptor variable in `data`, default is "cause" for
mortality data sets.}
}
\value{
a list in order containing minimum time period, maximum time period and cause name.
}
\description{
Get starting time period, ending time period and mortality cause name from the
data set for use in axis and main plot titles.
}
