% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_availability}
\alias{iidda_plot_availability}
\title{Plot Data Availability Graph}
\usage{
iidda_plot_availability(
  data,
  page,
  page_size,
  scale_colour,
  title_colour,
  title_totals,
  text_size = 9,
  left_margin = 100,
  legend_margin = 15,
  subplot_widths = c(5, 1)
)
}
\description{
Plot Data Availability Graph
}
