% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_box}
\alias{iidda_plot_box}
\title{Plot Box Plot}
\usage{
iidda_plot_box(
  plot_object,
  data = NULL,
  series_variable = "deaths",
  time_unit = "week",
  ...
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing data prepped for box plotting, typically output from `iidda_prep_box()`.
If `NULL` data is inherited from `plot_object`}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_unit}{time unit to display box plots on the x-axis. Defaults to "week", should be able to
handle any time_unit from iidda.analysis:::time_units.}

\item{...}{other arguments to be passed to `scale_x_discrete`}
}
\value{
a ggplot2 plot object containing a box plots of time series data
}
\description{
Add a box plot to an exiting ggplot plot object.  Graphical choices were made to closely reflect
plots generated with `LBoM::monthly_box_plot`.
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
