% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_heatmap_decomp}
\alias{iidda_plot_heatmap_decomp}
\title{IIDDA Heatmap Decomposition}
\usage{
iidda_plot_heatmap_decomp(
  plot_object,
  data = NULL,
  grouping_variable,
  series_variable,
  time_variable = "period_mid_time",
  num_days_variable = "num_days",
  contiguous_variable = NULL,
  trans = scales::sqrt_trans(),
  n_colours = (scales::brewer_pal(palette = "YlOrRd"))(9),
  NA_colour = "black",
  period_aggregator = PeriodAggregator(),
  layout_proportions = c(1, 3),
  rate_title = "",
  grouping_title = ""
)
}
\arguments{
\item{data}{Named list of two data frames: `heatmap` for the heatmap
component and `line` for the line-graph component. This list can be
computed using \code{\link{iidda_prep_heatmap_decomp}}}
}
\description{
IIDDA Heatmap Decomposition
}
