% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_highlight}
\alias{iidda_plot_highlight}
\title{Add Plot Highlight}
\usage{
iidda_plot_highlight(
  plot_object,
  data = NULL,
  series_variable = "deaths",
  time_variable = "period_start_date",
  filter_variable = "period_start_date",
  filter_start = "1700-01-01",
  filter_end = "1800-01-01",
  ...
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing time series data. If `NULL` data is inherited from `plot_object`. This
has only been tested with data output from `iidda_plot_ma`.}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_start_date"}

\item{filter_variable}{column name of variable to filter on in `data`, default is "period_start_date"}

\item{filter_start}{value of `filter_variable` for starting range, default is "1700-01-01"}

\item{filter_end}{value of `filter_variable` for ending range, default is "1800-01-01"}

\item{...}{other arguments to be passed to `ggforce::geom_mark_rect`, for example annotating with text}
}
\value{
a ggplot2 plot object a rectangular plot highlight
}
\description{
Add a rectangular highlighted region to an existing ggplot2 plot object
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_periodogram}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
