% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_ma}
\alias{iidda_plot_ma}
\title{Plot Moving Average Time Series}
\usage{
iidda_plot_ma(
  plot_object,
  data = NULL,
  series_variable = NULL,
  time_variable = NULL
)
}
\arguments{
\item{plot_object}{a \code{ggplot2} plot object}

\item{data}{data frame containing moving average time series data, typically output from \code{iidda_prep_ma()}.
If \code{NULL} data is inherited from \code{plot_object}}

\item{series_variable}{column name of series variable in \code{data}, default is "deaths"}

\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}
}
\value{
a ggplot2 plot object containing a moving average time series
}
\description{
Add a moving average time series line to an exiting ggplot plot object. Graphical choices were made
to closely reflect plots generated with \code{LBoM::plot.LBoM}.
}
\concept{plotting_functions}
