% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_periodogram}
\alias{iidda_plot_periodogram}
\title{Plot Periodogram}
\usage{
iidda_plot_periodogram(
  plot_object,
  data = NULL,
  period_variable = "per",
  spectral_density_variable = "spec",
  x_variable_name = "Period",
  y_variable_name = "Spectral Density",
  max_period = 10
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing time series data. If `NULL` data is inherited from `plot_object`.}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_start_date"}

\item{filter_variable}{column name of variable to filter on in `data`, default is "period_start_date"}

\item{filter_start}{value of `filter_variable` for starting range, default is "1700-01-01"}

\item{filter_end}{value of `filter_variable` for ending range, default is "1800-01-01"}

\item{...}{other arguments to be passed to `ggforce::geom_mark_rect`, for example annotating with text}
}
\value{
a ggplot2 plot object a rectangular plot highlight
}
\description{
Add a rectangular highlighted region to an existing ggplot2 plot object
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
