% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_scales}
\alias{iidda_plot_scales}
\title{Scale series data}
\usage{
iidda_plot_scales(plot_object, data = NULL, scale_transform = "log1p")
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing time series data. If `NULL` data is inherited from `plot_object`.}

\item{scale_transform}{transformation to apply to \code{y} variable, must be a valid ggplot2 transformation.}
}
\value{
a ggplot2 plot object with scaled \code{y} data
}
\description{
Scale time series data by transformation.
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_periodogram}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
