% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_series}
\alias{iidda_plot_series}
\title{Plot  Time Series}
\usage{
iidda_plot_series(
  plot_object,
  data = NULL,
  series_variable = "deaths",
  time_variable = "period_end_date",
  time_unit = "year"
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing time series data, typically output from `iidda_prep_series()`.
If `NULL` data is inherited from `plot_object`}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{time_unit}{time unit to display on the x-axis.}
}
\value{
a ggplot2 plot object containing a moving average time series
}
\description{
Add a time series line to an exiting ggplot plot object.
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
