% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_settings}
\alias{iidda_plot_settings}
\title{LBoM plot settings}
\usage{
iidda_plot_settings(
  plot_object,
  data,
  min_time = "min_time",
  max_time = "max_time",
  descriptor_name = "descriptor_name",
  theme = iidda_theme
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{list containing metadata. If `NULL` data is inherited from `plot_object`.}

\item{min_time}{name of field in data containing the minimum time period range, defaults to "min_time".}

\item{max_time}{name of field in data containing the minimum time period range, defaults to "max_time".}

\item{descriptor_name}{either the name of a field in data containing the descriptor or a string to be used as the
plot title. If there are too more than 3 elements in the descriptor field, then `descriptor_variable` is used as the plot
title.}

\item{theme}{ggplot theme}
}
\value{
a ggplot2 plot object with title, subtitle and adjusted theme.
}
\description{
Add basic features to a ggplot2 plot object including title, subtitle and classic `ggplot2::theme_bw` theme.
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
