% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_wavelet}
\alias{iidda_plot_wavelet}
\title{Plot Wavelet}
\usage{
iidda_plot_wavelet(
  plot_object,
  data = NULL,
  wavelet_data,
  contour_data,
  y_variable_name = "Period (years)",
  fill_variable_name = "Power",
  max_period = 10,
  colour_levels = 250,
  start_hue = 0,
  end_hue = 0.7,
  sig_lvl = 0.05
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing wavelet data prepped for use in \code{ggplot2::geom_tile}.
The output from \code{iidda_prep_wavelet} produces a data set prepped for this argument, named \code{tile_data_to_plot} in the returned list.
If `NULL` data is inherited from `plot_object`.}

\item{wavelet_data}{list containing raw wavelet transformed data, typically output from \code{WaveletComp::analyze.wavelet}.
The output from \code{iidda_prep_wavelet} produces a data set prepped for this argument, named \code{transformed_data} in the returned list.}

\item{contour_data}{data set containing contour data prepped for use in \code{ggplot2::geom_contour}.
The output from \code{iidda_prep_wavelet} produces a data set prepped for this argument, named \code{cont_data_to_plot} in the returned list.}

\item{y_variable_name}{name of y variable in plot, defaults to "Period (years)".}

\item{fill_variable_name}{name of colour fill variable in plot, defaults to "Power".}

\item{max_period}{maximum period to appear on the plot, defaults to 10 years.}

\item{colour_levels}{number of colours to pass to \code{scale_fill_gradientn}.}

\item{start_hue}{starting hue colour to pass to \code{scale_fill_gradientn}, default taken from WaveletComp::wt.image.}

\item{end_hue}{ending hue colour to pass to \code{scale_fill_gradientn}, default taken from WaveletComp::wt.image.}

\item{sig_lvl}{significance level for white contours}
}
\value{
a ggplot2 object of a wavelet
}
\description{
Plot wavelet to look similar to base R plot of \code{WaveletComp::wt.image} using ggplot2 functionality.
Some visual choices were made to reflect work done by Steven Lee (https://github.com/davidearn/StevenLee)
and Kevin Zhao (https://github.com/davidearn/KevinZhao).
}
\concept{plotting_functions}
