% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_heatmap}
\alias{iidda_prep_heatmap}
\title{Prep Data for Heatmap}
\usage{
iidda_prep_heatmap(
  data,
  series_variable = "deaths",
  start_time_variable = "period_start_date",
  end_time_variable = "period_end_date",
  time_unit = c("yday", "year"),
  prepend_string = "End ",
  normalize = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{start_time_variable}{column name of time variable in `data`, default is "period_start_date"}

\item{end_time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{time_unit}{a vector of new time unit fields to create from `start_time_variable` and `end_time_variable`.
Defaults to "c("yday","year")". The currently functionality expects that both "yday" and "year" are included, should be
made more general to incorporate any of plot_vars.R time_units.}

\item{prepend_string}{string to prepend to newly created time_unit fields to distinguish between time_unit
fields corresponding to starting versus ending time periods. Defaults to "End ". For example, a `time_unit` of "year"
will create a field name "Year" from `start_time_variable` and a field called "End Year" created from `end_time_variable`.}

\item{normalize}{boolean flag to normalize `series_variable` data to be between 0 and 1.}

\item{...}{optional arguments to `year_end_fix()`}
}
\value{
all fields in`data` with records prepped for plotting heatmaps. The name
of the new `time_unit` fields will be named from lubridate_funcs.
}
\description{
Prep data for heatmap plots. Prep steps were taken from `LBoM::seasonal_heat_map`
and they include creating additional time unit fields, splitting weeks that cover the
year end,  and optionally normalizing series data to be in the range (0,1).
}
\seealso{
Other prep_data_for_plotting: 
\code{\link{iidda_prep_bar}()},
\code{\link{iidda_prep_box}()},
\code{\link{iidda_prep_ma}()},
\code{\link{iidda_prep_periodogram}()},
\code{\link{iidda_prep_rohani}()},
\code{\link{iidda_prep_series}()},
\code{\link{iidda_prep_wavelet}()}
}
\concept{prep_data_for_plotting}
