% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_ma}
\alias{iidda_prep_ma}
\title{Prep Data for Moving Average Plot}
\usage{
iidda_prep_ma(
  data,
  series_variable = "deaths",
  time_variable = "period_end_date",
  trim_zeroes = TRUE,
  trim_series = TrimSeries(zero_lead = FALSE, zero_trail = FALSE),
  handle_missing_values = HandleMissingValues(na_remove = FALSE, na_replace = NULL),
  handle_zero_values = HandleZeroValues(zero_remove = FALSE, zero_replace = NULL),
  compute_moving_average = ComputeMovingAverage(ma_window_length = 52)
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{trim_zeroes}{boolean value to filter data to exclude leading and trailing zeroes}

\item{trim_series}{function to trim leading and trailing series zeroes, defaults to TrimSeries}

\item{handle_missing_values}{function to handle missing values, defaults to HandleMissingValues}

\item{handle_zero_values}{function to handle zero values, defaults to HandleZeroValues}

\item{compute_moving_average}{function to compute the moving average of `series_variable`}
}
\value{
all fields in `data` with records prepped for plotting moving average time series
}
\description{
Prep data for plotting moving average. Prep steps were taken from `LBoM::plot.LBoM` and they include
handling missing values and zeroes, optionally trimming time series and computing the moving
average.
}
\seealso{
Other prep_data_for_plotting: 
\code{\link{iidda_prep_bar}()},
\code{\link{iidda_prep_box}()},
\code{\link{iidda_prep_heatmap}()},
\code{\link{iidda_prep_periodogram}()},
\code{\link{iidda_prep_rohani}()},
\code{\link{iidda_prep_series}()},
\code{\link{iidda_prep_wavelet}()}
}
\concept{prep_data_for_plotting}
