% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_periodogram}
\alias{iidda_prep_periodogram}
\title{Prep Data for Periodogram}
\usage{
iidda_prep_periodogram(
  data,
  series_variable = "deaths",
  time_variable = "period_end_date",
  transform = FALSE,
  transformation = "log10",
  spans = NULL,
  kernel = NULL,
  taper = 0.1,
  pad = 0,
  fast = TRUE,
  demean = FALSE,
  detrend = TRUE,
  na.action = na.fail,
  normalize = TRUE,
  periods_per_year = 52,
  handle_missing_values = HandleMissingValues(na_remove = TRUE, na_replace = NULL)
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{handle_missing_values}{function to handle missing values, defaults to HandleMissingValues}

\item{time_unit}{time unit to create field from `time_variable`. Must be one of iidda.analysis:::time_units, defaults to "week".}

\item{handle_zero_values}{function to handle zero values, defaults to HandleZeroValues}
}
\value{
all fields in`data` with records prepped for plotting box plots. The name
of the new `time_unit` field will be named from lubridate_funcs.
}
\description{
Prep data for plotting peridograms. Prep steps were taken from `LBoM::periodogram`
and they include handling missing values,...
}
\seealso{
Other prep_data_for_plotting: 
\code{\link{iidda_prep_bar}()},
\code{\link{iidda_prep_box}()},
\code{\link{iidda_prep_heatmap}()},
\code{\link{iidda_prep_ma}()},
\code{\link{iidda_prep_rohani}()},
\code{\link{iidda_prep_series}()},
\code{\link{iidda_prep_wavelet}()}
}
\concept{prep_data_for_plotting}
