% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_rohani}
\alias{iidda_prep_rohani}
\title{Prep Data for Rohani Plot}
\usage{
iidda_prep_rohani(
  data,
  series_variable = NULL,
  time_variable = "period_end_date",
  start_time_variable = "period_start_date",
  time_unit = "year",
  grouping_variable = "cause",
  ranking_variable = NULL,
  normalize = FALSE,
  handle_missing_values = HandleMissingValues(na_remove = FALSE, na_replace = NULL),
  handle_zero_values = HandleZeroValues(zero_remove = FALSE, zero_replace = NULL),
  create_nonexistent = FALSE
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{series_variable}{column name of series variable in \code{data}, default is "deaths"}

\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{start_time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{time_unit}{a vector of new time unit fields to create from \code{start_time_variable} and \code{end_time_variable}.
Defaults to "c("year")". The currently functionality expects that "year" is included, should be
made more general to incorporate any of iidda.analysis:::time_units.}

\item{grouping_variable}{column name of grouping variable to appear on the y-axis of the heatmap.}

\item{ranking_variable}{column name of variable used to rank the grouping variable.}

\item{normalize}{boolean flag to normalize \code{series_variable} data to be between 0 and 1.}

\item{handle_missing_values}{function to handle missing values, defaults to HandleMissingValues}

\item{handle_zero_values}{function to handle zero values, defaults to HandleZeroValues}

\item{create_nonexistent}{boolean flag to create \code{NA} records for non-existent \code{time_unit} and \code{grouping_variable}.
This creates all combinations of \code{time_unit} and \code{grouping_variable} to ensure there are no missing records.}
}
\value{
all fields in\code{data} with records prepped for plotting rohani heatmaps. The name
of the new \code{time_unit} fields will be named from lubridate_funcs.
}
\description{
Prep data for rohani plots. Prep steps include creating additional time unit fields, summarizing the series
variable by time unit and grouping variable (the x and y axis variables) ,and optionally normalizing series
data to be in the range (0,1). By default, the grouping variable is ranked in order of the summarized series
variable. Needs to be generalized more, might
need to handle the case where the desired y-axis is a second time unit, as in the seasonal heatmap plot and therefore
making use of the year_end_fix function.
}
\concept{prep_data_for_plotting}
