% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_series}
\alias{iidda_prep_series}
\title{Prep Data for Time Series}
\usage{
iidda_prep_series(
  data,
  series_variable = "deaths",
  time_variable = "period_end_date",
  grouping_variable = "cause",
  time_unit = "year",
  summarize_series = TRUE,
  trim_zeroes = TRUE,
  trim_series = TrimSeries(zero_lead = FALSE, zero_trail = FALSE),
  handle_missing_values = HandleMissingValues(na_remove = FALSE, na_replace = NULL),
  handle_zero_values = HandleZeroValues(zero_remove = FALSE, zero_replace = NULL)
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{grouping_variable}{column name of the grouping variable in `data` to summarize the series variable over,
if `summarize=TRUE`}

\item{time_unit}{time unit to sum series data over, must be one of iidda.analysis:::time_units, defaults to "year".}

\item{summarize_series}{boolean value to indicate summarizing by `time_unit` over the series variable}

\item{trim_zeroes}{boolean value to filter data to exclude leading and trailing zeroes}

\item{trim_series}{function to trim leading and trailing series zeroes, defaults to TrimSeries}

\item{handle_missing_values}{function to handle missing values, defaults to HandleMissingValues}

\item{handle_zero_values}{function to handle zero values, defaults to HandleZeroValues}
}
\value{
all fields in `data` with records prepped for plotting moving average time series
}
\description{
Prep data for basic time series plot. Prep steps were taken from `LBoM::plot.LBoM` and they include
handling missing values and zeroes, and optionally trimming time series.
}
\seealso{
Other prep_data_for_plotting: 
\code{\link{iidda_prep_bar}()},
\code{\link{iidda_prep_box}()},
\code{\link{iidda_prep_heatmap}()},
\code{\link{iidda_prep_ma}()},
\code{\link{iidda_prep_periodogram}()},
\code{\link{iidda_prep_rohani}()},
\code{\link{iidda_prep_wavelet}()}
}
\concept{prep_data_for_plotting}
