% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_wavelet}
\alias{iidda_prep_wavelet}
\title{Prep Data for Wavelet Plot}
\usage{
iidda_prep_wavelet(
  data,
  trend_data,
  time_variable = "period_end_date",
  series_variable = "deaths",
  trend_variable = "deaths",
  series_suffix = "_series",
  trend_suffix = "_trend",
  wavelet_variable = "detrend_norm",
  output_emd_trend = "emd_trend",
  output_norm = "norm",
  output_sqrt_norm = "sqrt_norm",
  output_log_norm = "log_norm",
  output_emd_norm = "emd_norm",
  output_emd_sqrt = "emd_sqrt",
  output_emd_log = "emd_log",
  output_detrend_norm = "detrend_norm",
  output_detrend_sqrt = "detrend_sqrt",
  output_detrend_log = "detrend_log",
  data_harmonizer = SeriesHarmonizer(time_variable, series_variable),
  trend_data_harmonizer = SeriesHarmonizer(time_variable, trend_variable),
  data_deheaper = WaveletDeheaper(time_variable, series_variable),
  trend_deheaper = WaveletDeheaper(time_variable, trend_variable),
  joiner = WaveletJoiner(time_variable, series_suffix, trend_suffix),
  interpolator = WaveletInterpolator(time_variable, series_variable, trend_variable,
    series_suffix, trend_suffix),
  normalizer = WaveletNormalizer(time_variable, series_variable, trend_variable,
    series_suffix, trend_suffix, output_emd_trend, output_norm, output_sqrt_norm,
    output_log_norm, output_emd_norm, output_emd_sqrt, output_emd_log,
    output_detrend_norm, output_detrend_sqrt, output_detrend_log),
  transformer = WaveletTransformer(time_variable, wavelet_variable, dt = 1/52, dj = 1/50,
    lowerPeriod = 1/2, upperPeriod = 10, n.sim = 1000, make.pval = TRUE, date.format =
    "\%Y-\%m-\%d")
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{trend_data}{data frame containing time series trend data}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{series_variable}{column name of series variable in `data`, default is "deaths_series"}

\item{trend_variable}{column name of series variable in `data`, default is "deaths_trend"}

\item{series_suffix}{suffix to be appended to series data fields}

\item{trend_suffix}{suffix to be appended to trend data fields}

\item{wavelet_variable}{name of the field in `data` to be wavelet transformed}

\item{output_emd_trend}{name of output field for the empirical mode decomposition applied to `trend_variable`}

\item{output_norm}{name of output field for the `series_variable` normalized by `output_emd_trend`}

\item{output_sqrt_norm}{name of output field for the square root of `output_norm`}

\item{output_log_norm}{name of output field for the logarithm of (`output_norm` + `eps`)}

\item{output_emd_norm}{name of output field for the empirical mode decomposition applied to `output_norm`}

\item{output_emd_sqrt}{name of output field for the empirical mode decomposition applied to `output_sqrt_norm`}

\item{output_emd_log}{name of output field for the empirical mode decomposition applied to `output_log_norm`}

\item{output_detrend_norm}{name of output field for the computed field `output_norm`-`output_emd_norm`}

\item{output_detrend_sqrt}{name of output field for the computed field `output_sqrt_norm`-`output_emd_sqrt`}

\item{output_detrend_log}{name of output field for the computed field `output_log_norm`-`output_emd_log`}

\item{data_harmonizer}{function that harmonizes time scales and series names so there is one
data point per time unit}

\item{trend_data_harmonizer}{function that harmonizes time scales and trend names so there is one
data point per time unit}

\item{data_deheaper}{function that fixes heaping errors on series data}

\item{trend_deheaper}{function that fixes heaping errors on trend data}

\item{joiner}{function that joins series and trend data sets}

\item{interpolator}{function that linearly interpolates series and trend data}

\item{normalizer}{function that computes normalized fields}

\item{transformer}{function that computes wavelet transform}
}
\value{
list containing:
       * \code{transforemd_data} - wavelet transformed data
       * \code{tile_data_to_plot} - data set of the wavelet transformed data
prepped for plotting with \code{ggplot2::geom_tile}
       * \code{contour_data_to_plot} - data set of the transformed wavelet data prepped
for plotting with \code{ggplot2::geom_contour}
}
\description{
Prep data for wavelet plot. Prep steps were taken from code provided by Steven Lee
(https://github.com/davidearn/StevenLee) and Kevin Zhao (https://github.com/davidearn/KevinZhao).
}
\seealso{
Other prep_data_for_plotting: 
\code{\link{iidda_prep_bar}()},
\code{\link{iidda_prep_box}()},
\code{\link{iidda_prep_ma}()},
\code{\link{iidda_prep_rohani}()},
\code{\link{iidda_prep_seasonal_heatmap}()},
\code{\link{iidda_prep_series}()}
}
\concept{prep_data_for_plotting}
