% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-plot.R
\name{iidda_render_plot}
\alias{iidda_render_plot}
\title{Render a Plot}
\usage{
iidda_render_plot(
  data,
  title = TitleGuesser(),
  subtitle = TimeRangeDescriber(),
  theme = iidda_theme
)
}
\arguments{
\item{data}{A data frame with an attached plot object.}

\item{title}{A plot title, or a function that takes \code{data} and returns a
title string.}

\item{subtitle}{A plot subtitle, or a function that takes \code{data} and returns
a subtitle string.}

\item{theme}{A function that returns a \code{ggplot2::theme} object.}
}
\description{
Extracts a plot object attached to a data set, applies a title, subtitle,
and theme, and returns the plot.
}
