% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{iidda_theme}
\alias{iidda_theme}
\alias{iidda_theme_time}
\alias{iidda_theme_heat}
\alias{iidda_theme_above}
\title{Themes for ggplot2}
\usage{
iidda_theme()

iidda_theme_time()

iidda_theme_heat()

iidda_theme_above()
}
\description{
Themes for ggplot2
}
\section{Functions}{
\itemize{
\item \code{iidda_theme_time()}: Theme for plots where the x-axis represents time.
No x-axis titles will be plotted with this theme, because the meaning of a
time axis is obvious.

\item \code{iidda_theme_heat()}: Theme for heatmaps where the x-axis represents time.
No x-axis titles will be plotted with this theme, because the
meaning of a time axis is obvious. Grid lines are not plotted with this
theme because interpretation can be compromised when grid lines
are visible through the colours of the heatmap.

\item \code{iidda_theme_above()}: Theme for plots where the x-axis represents time,
but for which time information is not displayed because there are vertically
aligned plots below with the same time axis.

}}
