% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{is_leaf_disease}
\alias{is_leaf_disease}
\title{Is Leaf Disease}
\usage{
is_leaf_disease(disease, nesting_disease)
}
\arguments{
\item{disease}{Disease name vector.}

\item{nesting_disease}{Vector of the same length as \code{disease} giving
the nesting diseases of element in \code{disease}.}
}
\value{
True if disease is never a nesting disease (it is a leaf disease),
False if disease is a nesting disease.
}
\description{
Given a set of `disease`-`nesting_disease` pairs that all share the same
\code{\link{basal_disease}},
}
