% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{join_lookup_table}
\alias{join_lookup_table}
\title{Join lookup table}
\usage{
join_lookup_table(raw_data, lookup_type)
}
\arguments{
\item{raw_data}{data frame of table to be harmonized}

\item{lookup_type}{string indicating type of lookup table from API to join}
}
\value{
data frame of harmonized data with keys from API
}
\description{
Joins lookup table in API to data
}
