% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{join_user_table}
\alias{join_user_table}
\title{Join user-defined lookup table}
\usage{
join_user_table(raw_data, user_table_path, lookup_type, join_by = c())
}
\arguments{
\item{raw_data}{data frame of table to be harmonized}

\item{user_table_path}{string indicating path to user-defined lookup table}

\item{lookup_type}{string indicating type of lookup table (disease, location, sex). Used to determine columns to join by if \code{join_by} not specified}

\item{join_by}{vector of strings indicating columns to join by (optional if \code{lookup_type} is disease, location, or sex)}
}
\value{
data frame of harmonized data with user-defined keys
}
\description{
Joins user-defined lookup table to data
}
