% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{log1p_modified_trans}
\alias{log1p_modified_trans}
\title{Log1p Scale Transformation}
\usage{
log1p_modified_trans(n = 10)
}
\arguments{
\item{n}{number of desired breaks}
}
\value{
a \code{scales::trans_new} function
}
\description{
Slight modification of \code{log1p_trans()} to include better breaks that are log1p-based (log-based and
shifted 1 so that breaks can be computed in the presence of zeroes.)
}
