% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\docType{data}
\name{lubridate_funcs}
\alias{lubridate_funcs}
\title{Lubridate functions}
\format{
An object of class \code{character} of length 10.
}
\usage{
lubridate_funcs
}
\description{
lubridate functions with desired interpretable labels
}
\keyword{datasets}
