% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{make_time_trans}
\alias{make_time_trans}
\title{Get time transformation}
\usage{
make_time_trans(unit = unname(time_units))
}
\arguments{
\item{unit}{time unit, one or more of iidda.analysis:::time_units}
}
\value{
function to compute time unit
}
\description{
Get associated lubridate function to compute time unit.
}
