% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{mutate_time_vars}
\alias{mutate_time_vars}
\title{Mutate time variables}
\usage{
mutate_time_vars(
  data,
  unit = unname(time_units),
  input_nm = "period_end_date",
  output_nm = get_unit_labels(unit)
)
}
\arguments{
\item{data}{data set containing an input time field}

\item{unit}{time unit, one of time_units}

\item{input_name}{field name in `data` containing input time field}

\item{output_name}{field name of newly created time unit field, by default uses get_unit_labels().}
}
\value{
all fields in `data` with additional time unit field
}
\description{
Create new time unit fields
}
