% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{names_to_join_by}
\alias{names_to_join_by}
\title{Column names to join by}
\usage{
names_to_join_by(lookup_type)
}
\arguments{
\item{lookup_type}{string indicating type of lookup table (disease, location, sex, age group)}
}
\value{
vector of column names to join by for the type of lookup table
}
\description{
Defines column names to join by for a type of lookup table
}
