% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{normalize_disease_hierarchy}
\alias{normalize_disease_hierarchy}
\title{Normalize Disease Hierarchy}
\usage{
normalize_disease_hierarchy(
  data,
  disease_lookup,
  grouping_columns = c("period_start_date", "period_end_date", "location"),
  basal_diseases_to_prune = character(),
  find_unaccounted_cases = TRUE,
  specials_pattern = "_unaccounted$"
)
}
\arguments{
\item{data}{A tidy data set with the following minimal set of columns:
\code{disease}, \code{nesting_disease}, \code{basal_disease}, \code{period_start_date},
\code{period_end_date}, and \code{location}. Note that the latter three can be
modified with \code{grouping_columns}.}

\item{disease_lookup}{A lookup table with \code{disease} and \code{nesting_disease}
columns that describe a global disease hierarchy that will be applied
locally to flatten disease hierarchy at each point in time and space
in the tidy data set in the \code{data} argument.}

\item{grouping_columns}{Character vector of column names to use when
grouping to determine the context.}

\item{basal_diseases_to_prune}{Character vector of \code{disease}s to
remove from \code{data}.}

\item{find_unaccounted_cases}{Make new records for instances when the sum of
leaf diseases is less than the reported total for their basal disease.}

\item{specials_pattern}{Optional regular expression to use to match
\code{disease} names in \code{data} that should be added to the lookup table. This
is useful for disease names that are not historical and produced for
harmonization purposes. The most common example is \code{"_unaccounted$"},
which is the default. Setting this argument to \code{NULL} avoids adding
any special disease names to the lookup table.}
}
\description{
Take a tidy data set with a potentially complex disease hierarchy
and flatten this hierarchy so that, at any particular time and location
(or some other context), all diseases in the \code{disease} column have the
same \code{nesting_disease}.
}
\concept{normalization}
