% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{normalize_location}
\alias{normalize_location}
\title{Normalize Location}
\usage{
normalize_location(data)
}
\arguments{
\item{data}{Tidy dataset with an iso_3166_2 column}
}
\value{
Tidy dataset without country-level data and with provinces and 
territories geographically ordered
}
\description{
Set geographic order of provinces and territories and remove country-level
data
}
