% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{normalize_population}
\alias{normalize_population}
\title{Normalize Population}
\usage{
normalize_population(data, harmonized_population)
}
\arguments{
\item{data}{Tidy dataset with columns period_start_date, period_end_date
iso_3166_2.}

\item{harmonized_population}{Harmonized population data with
columns date, iso_3166_2, and population (other columns will be dropped).}
}
\value{
Tidy dataset joined with harmonized population.
}
\description{
Normalize Population
}
\concept{normalization}
