% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_tools.R
\name{normalize_time_scales}
\alias{normalize_time_scales}
\title{Normalize Time Scales}
\usage{
normalize_time_scales(
  data,
  initial_group = c("year", "iso_3166", "iso_3166_2", "disease", "nesting_disease",
    "basal_disease"),
  final_group = c("basal_disease"),
  get_implied_zeros = TRUE,
  aggregate_if_unavailable = TRUE
)
}
\arguments{
\item{data}{A tidy data set with a `time_scale` and `year` column}

\item{initial_group}{Character vector naming columns for defining
the initial grouping used to compute the shortest time scales.}

\item{final_group}{Character vector naming columns for defining the final
grouping used to compute the longest of the shortest time scales.}

\item{get_implied_zeros}{Add zeros that are implied by a '0' reported at a coarser timescale.}

\item{aggregate_if_unavailable}{If a location is not reporting for the determined
'best timescale', but is reporting at a finer timescale, aggregate this finer
timescale to the 'best timescale'}
}
\value{
A data set only containing records with the optimal time scale.
}
\description{
Choose a single best `time_scale` for each year in a dataset, grouped by
nesting disease. This best `time_scale` is defined as the longest
of the shortest time scales in each location and sub-disease.
}
