% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_averager.R
\name{num_days}
\alias{num_days}
\alias{num_days_util}
\title{Numbers of Days}
\usage{
num_days(start_date, end_date)

num_days_util(start_date, end_date)
}
\arguments{
\item{start_date}{Vector of period starting dates}

\item{end_date}{Vector of period ending dates}
}
\description{
Compute a vector giving the number of days in a
set of periods, given equal length vectors of the start date and end date of
these periods. This
}
\section{Functions}{
\itemize{
\item \code{num_days_util()}: Low-level interface for `num_days`.

}}
\seealso{
Other time_periods: 
\code{\link{mid_dates_times}},
\code{\link{period_averager}()}
}
\concept{time_periods}
