% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_averager.R
\name{period_averager}
\alias{period_averager}
\title{Obtain period midpoints and average daily rates for count data}
\usage{
period_averager(
  data,
  count_col = "cases_this_period",
  start_col = "period_start_date",
  end_col = "period_end_date",
  count_name = NULL,
  keep_raw = TRUE,
  keep_cols = c(count_name)
)
}
\arguments{
\item{data}{Data frame with rows at minimum containing period start and end
dates and a count variable.}

\item{count_col}{Character, name of count data column.}

\item{start_col}{Character, name of start date column.}

\item{end_col}{Character, name of end date column.}

\item{count_name}{Character, name of column specifying counted variable.}

\item{keep_raw}{Logical, retain *_col columns in output.}

\item{keep_cols}{Character vector, names of coluimns to retain in output.
To get all columns set \code{keep_cols} to \code{NULL}.}
}
\value{
Data frame containing period length (\code{num_days}),
period mid-date (\code{period_mid_date}),
average daily count (\code{daily_rate}), and any additional data from `data`.
}
\description{
Obtain period midpoints and average daily rates for count data
}
\examples{
set.seed(666)
data <- data.frame(disease = "senioritis"
 , period_start_date = seq(as.Date("2023-04-03"), as.Date("2023-06-05"), by = 7)
 , period_end_date = seq(as.Date("2023-04-09"), as.Date("2023-06-11"), by = 7)
 , cases_this_period = sample(0:100, 10, replace = TRUE)
 , location = "college"
)

period_averager(data, keep_raw = TRUE, keep_cols = c("disease", "location"))

}
\seealso{
Other time_periods: 
\code{\link{mid_dates_times}},
\code{\link{num_days}()}
}
\concept{time_periods}
