% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{PeriodAggregator}
\alias{PeriodAggregator}
\title{Period Aggregator}
\usage{
PeriodAggregator(rate_variable, norm_exponent = 5)
}
\arguments{
\item{rate_variable}{Name of variable to be used to store the normalized
count variable.}

\item{norm_exponent}{Exponent to use in normalization. The default is \code{5},
which means \verb{per 100,000}.}
}
\value{
A function like \code{\link{period_aggregator_default}} that
aggregates data so that each time period is represented by exactly
one record.
}
\description{
Create function that aggregates information over time periods, normalizes
a count variable, and creates new fields to summarize this information.
}
\concept{data_prep_constructors}
