% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{quantile_trans}
\alias{quantile_trans}
\title{Quantile Transformation}
\usage{
quantile_trans(x)
}
\arguments{
\item{x}{vector to be transformed}
}
\value{
a \code{scales::trans_new} function
}
\description{
Quantile transformation, adapted from
https://stackoverflow.com/questions/38874741/transform-color-scale-to-probability-transformed-color-distribution-with-scale-f
}
