% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_iidda_dataset}
\alias{read_iidda_dataset}
\title{Read IIDDA Dataset into a Dataframe}
\usage{
read_iidda_dataset(dataset_id)
}
\arguments{
\item{dataset_id}{ID for a dataset in the IIDDA}
}
\description{
Read IIDDA Dataset into a Dataframe
}
