% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_time_plots.R
\name{ref}
\alias{ref}
\title{Tag Reference}
\usage{
ref(x)
}
\arguments{
\item{x}{an object in a group}
}
\value{
the input object with the attribute
\code{'iidda_time_plot_reference'} equal to \code{TRUE}
}
\description{
Identify an object in a group of objects to be used to
determine aspects of the group.
}
