% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_tools.R
\name{resolve_join}
\alias{resolve_join}
\title{Resolve left_join}
\usage{
resolve_join(df)
}
\arguments{
\item{df}{data frame with duplicate columns ending in \code{.x} and \code{.y}}
}
\value{
data frame with one remaining column for duplicates
}
\description{
Resolves any duplicate columns that results after left_join due to shared columns between data frames.
Rule: Keeps old values if all newly joined values are NA. Keeps new values otherwise (even if some entries are empty)
}
\concept{harmonization}
