% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{SeriesHarmonizer}
\alias{SeriesHarmonizer}
\title{Series Harmonizer}
\usage{
SeriesHarmonizer(sum_fn = base::sum)
}
\arguments{
\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{series_variable}{column name of series variable in \code{data}, default is "deaths"}
}
\value{
A function like \code{\link{series_harmonizer_default}} to
harmonize disease/cause names.
}
\description{
Harmonizes the series variable in \code{data} so there is one data value for each time
unit in time variable (to account for different variations in disease/cause name)
}
\concept{data_prep_constructors}
