% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_time_plots.R
\name{shared_time_limits}
\alias{shared_time_limits}
\title{Shared Time Limits}
\usage{
shared_time_limits(..., time_id)
}
\arguments{
\item{...}{list of objects}

\item{time_id}{identifier for finding time axis information
in each object}
}
\value{
Date vector of the limits of a shared time axis
}
\description{
Get limits of a time-axis that is shared by several
objects.
}
