% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std_series_variables}
\alias{std_series_variables}
\title{Series variables}
\usage{
std_series_variables()
}
\description{
List of column names that contain numerical information that could
represent a time series if processed appropriately.
}
