% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_time_plots.R
\name{time_extent}
\alias{time_extent}
\title{Time Extent}
\usage{
time_extent(x, time_id)
}
\arguments{
\item{x}{an object}

\item{time_id}{identifier for finding time axis information
in the object}
}
\description{
Length of time in days representated by an object
}
