% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_time_plots.R
\name{time_harmonize}
\alias{time_harmonize}
\title{Time Harmonize}
\usage{
time_harmonize(..., time_id)
}
\arguments{
\item{...}{list of objects}

\item{time_id}{identifier for finding time axis information
in each object}
}
\value{
modified list of objects
}
\description{
Harmonize the time-axis over a group of objects
}
