% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_scale_picker.R
\name{time_scale_picker}
\alias{time_scale_picker}
\title{Default Time Scale Picker}
\usage{
time_scale_picker(data, time_scale_variable = NULL, time_group_variable = NULL)
}
\arguments{
\item{data}{Data to transform.}
}
\description{
Default Time Scale Picker
}
