% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\docType{data}
\name{time_units}
\alias{time_units}
\title{Time units}
\format{
An object of class \code{character} of length 29.
}
\usage{
time_units
}
\description{
Vector of all possible time units, most or all are derived from lubridate functions
}
\keyword{datasets}
