% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{TimeScalePicker}
\alias{TimeScalePicker}
\title{Time Scale Picker}
\usage{
TimeScalePicker()
}
\arguments{
\item{time_scale_variable}{Variable identifying the time scale of records.
The values of such a variable should be things like \code{"wk"}, \code{"mo"}, \code{"yr"}.}

\item{time_group_variable}{Variable identifying a grouping variable for
the time scales (e.g. a column identifying the year.).}
}
\description{
Time Scale Picker
}
