% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{TimeVariableConverter}
\alias{TimeVariableConverter}
\title{Time Variable Converter}
\usage{
TimeVariableConverter(
  as_date = as.Date,
  as_integer = as.integer,
  as_numeric = as.numeric
)
}
\arguments{
\item{as_date}{Function that takes a vector and converts it to a date vector if possible. Used only if \code{time_variable} is in \code{std_date_variables()}.}

\item{as_integer}{Function that takes a vector and converts it to an integer vector if possible. Used only if \code{time_variable} is in \code{std_integer_variables()}.}

\item{as_numeric}{Function that takes a vector and converts it to a numeric vector if possible. Used only if \code{time_variable} is in \code{std_numeric_variables()}.
\subsection{Returned Function}{
\itemize{
\item Arguments :
\itemize{
\item \code{data} : Data frame containing a time variable.
\item \code{time_variable} : Column name of time variable in \code{data}. The default
is \code{"period_end_date"}.
}
\item Return : A version of \code{data} with \code{time_variable} column converted to
}
}}
}
\description{
Construct a function that takes a data frame and returns another data
frame with a time variable converted so that it has the correct format,
class, and/or type.
}
