% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{TitleGuesser}
\alias{TitleGuesser}
\title{Title Guesser}
\usage{
TitleGuesser(custom = NULL, prefer = std_title_variables())
}
\arguments{
\item{custom}{Custom string for the title}

\item{prefer}{List of variables that could contain title information
in an order that will be used to find variables that will be used to
guess at a title. The first variable found in the data is the one that
is chosen.}
}
\description{
Title Guesser
}
