% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titleize.R
\name{titleize}
\alias{titleize}
\title{Titleize}
\usage{
titleize(title_info, max_items = 3L, max_chars = 15L)
}
\arguments{
\item{title_info}{Character vector to be summarized into a title}

\item{max_items}{TODO}

\item{max_chars}{TODO}
}
\description{
Convert a character vector (i.e. a character column) into a title for
a plot.
}
