% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_query_tools.R
\name{unique_entries}
\alias{unique_entries}
\title{Get unique tokens from iidda metadata}
\usage{
unique_entries(entries, metadata_search)
}
\arguments{
\item{entries}{List returned by \code{iidda.api::ops_staging$metadata}}

\item{metadata_search}{Character, field from which unique tokens are desired}
}
\value{
Character vector of unique tokens for a given field from all iidda
datasets
}
\description{
Get unique tokens from iidda metadata
}
