% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess-variable-names.R
\name{valid_time_vars}
\alias{valid_time_vars}
\title{Validate time variables}
\usage{
valid_time_vars(var_nm, data)
}
\arguments{
\item{var_nm}{string of variable name}

\item{data}{data frame}
}
\value{
boolean of validation status
}
\description{
Validate if variable is a date data type in the data set.
}
