% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{wavelet_something}
\alias{wavelet_something}
\title{Prep Data for Wavelet Plot}
\arguments{
\item{data}{data frame containing time series data}

\item{trend_data}{data frame containing time series trend data}

\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{series_variable}{column name of series variable in \code{data}, default is "deaths_series"}

\item{trend_variable}{column name of series variable in \code{data}, default is "deaths_trend"}

\item{series_suffix}{suffix to be appended to series data fields}

\item{trend_suffix}{suffix to be appended to trend data fields}

\item{wavelet_variable}{name of the field in \code{data} to be wavelet transformed}

\item{output_emd_trend}{name of output field for the empirical mode decomposition applied to \code{trend_variable}}

\item{output_norm}{name of output field for the \code{series_variable} normalized by \code{output_emd_trend}}

\item{output_sqrt_norm}{name of output field for the square root of \code{output_norm}}

\item{output_log_norm}{name of output field for the logarithm of (\code{output_norm} + \code{eps})}

\item{output_emd_norm}{name of output field for the empirical mode decomposition applied to \code{output_norm}}

\item{output_emd_sqrt}{name of output field for the empirical mode decomposition applied to \code{output_sqrt_norm}}

\item{output_emd_log}{name of output field for the empirical mode decomposition applied to \code{output_log_norm}}

\item{output_detrend_norm}{name of output field for the computed field \code{output_norm}-\code{output_emd_norm}}

\item{output_detrend_sqrt}{name of output field for the computed field \code{output_sqrt_norm}-\code{output_emd_sqrt}}

\item{output_detrend_log}{name of output field for the computed field \code{output_log_norm}-\code{output_emd_log}}

\item{data_harmonizer}{function that harmonizes time scales and series names so there is one
data point per time unit}

\item{trend_data_harmonizer}{function that harmonizes time scales and trend names so there is one
data point per time unit}

\item{data_deheaper}{function that fixes heaping errors on series data}

\item{trend_deheaper}{function that fixes heaping errors on trend data}

\item{joiner}{function that joins series and trend data sets}

\item{interpolator}{function that linearly interpolates series and trend data}

\item{normalizer}{function that computes normalized fields}

\item{transformer}{function that computes wavelet transform}
}
\value{
list containing:
* \code{transforemd_data} - wavelet transformed data
* \code{tile_data_to_plot} - data set of the wavelet transformed data
prepped for plotting with \code{ggplot2::geom_tile}
* \code{contour_data_to_plot} - data set of the transformed wavelet data prepped
for plotting with \code{ggplot2::geom_contour}
}
\description{
Prep data for wavelet plot. Prep steps were taken from code provided by Steven Lee
(https://github.com/davidearn/StevenLee) and Kevin Zhao (https://github.com/davidearn/KevinZhao).
}
\concept{prep_data_for_plotting}
