% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{WaveletInterpolator}
\alias{WaveletInterpolator}
\title{Wavelet Interpolator}
\usage{
WaveletInterpolator(
  time_variable = "period_start_date",
  series_variable = "deaths",
  trend_variable = "deaths",
  series_suffix = "_series",
  trend_suffix = "_trend"
)
}
\arguments{
\item{time_variable}{column name of time variable in `data`, default is "period_start_date"}

\item{series_variable}{column name of series variable in `data`, default is "deaths_series"}

\item{trend_variable}{column name of series variable in `data`, default is "deaths_trend"}

\item{series_suffix}{suffix to be appended to series data fields}

\item{trend_suffix}{suffix to be appended to trend data fields}
}
\value{
function that linearly interpolates series and trend data.

## Returned Function

- Arguments
   * `data` data frame containing time series data
- Return - `data` with linearly interpolated `series_variable` and
`trend_variable`
}
\description{
Linearly interpolates `NA` values in both series and trend variables.
}
\seealso{
Other data_prep_constructors: 
\code{\link{ComputeMovingAverage}()},
\code{\link{HandleMissingValues}()},
\code{\link{HandleZeroValues}()},
\code{\link{SeriesHarmonizer}()},
\code{\link{TrimSeries}()},
\code{\link{WaveletDeheaper}()},
\code{\link{WaveletJoiner}()},
\code{\link{WaveletNormalizer}()},
\code{\link{WaveletTransformer}()},
\code{\link{data_prep_constructors}}
}
\concept{data_prep_constructors}
