% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{WaveletJoiner}
\alias{WaveletJoiner}
\title{Wavelet Joiner}
\usage{
WaveletJoiner(
  time_variable = "period_end_date",
  series_suffix = "_series",
  trend_suffix = "_trend",
  keep_series_dates = TRUE
)
}
\arguments{
\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{series_suffix}{suffix to be appended to series data fields}

\item{trend_suffix}{suffix to be appended to trend data fields}

\item{keep_series_dates}{boolean flag to indicate if the dates in \code{series_data} should
be kept and data from \code{trend_data} is left joined, if \code{FALSE} dates from \code{trend_data}
are left joined instead}
}
\value{
function to join data and trend data sets
\subsection{Returned Function}{
\itemize{
\item Arguments
\itemize{
\item \code{series_data} data frame containing time series data
\item \code{trend_data} data frame containing trend data
}
\item Return - joined data set by \code{time_variable} with updated field names
}
}
}
\description{
Joins series data and trend datasets and keeps all time units in one of the
datasets.
}
\concept{data_prep_constructors}
