% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{ComputeMovingAverage}
\alias{ComputeMovingAverage}
\title{Compute Moving Average of Time Series}
\usage{
ComputeMovingAverage(ma_window_length = 52)
}
\arguments{
\item{ma_window_length}{length of moving average window, this will depend on the time scale in the data.
Defaults to 52, so that weekly data is averaged over years.}
}
\value{
A function like \code{\link{compute_moving_average_default}} to
remove to compute the moving average of a time series variable.
}
\description{
Compute Moving Average of Time Series
}
\concept{data_prep_constructors}
