% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{HandleMissingValues}
\alias{HandleMissingValues}
\title{Handle Missing Values in Series Variable}
\usage{
HandleMissingValues(na_remove = FALSE, na_replace = NULL)
}
\arguments{
\item{na_remove}{boolean value, if `TRUE` remove `NA`s in series variable}

\item{na_replace}{numeric value to replace `NA`s in series variable, if NULL
no replacement is performed}
}
\value{
a function to remove or replace missing values.

## Returned Function

- Arguments
   * `data` data frame containing time series data
   * `series_variable` column name of series variable in `data`, default is "deaths"
- Return - all fields in `data` with either `NA` records removed or replaced
}
\description{
Remove or replace values that are \code{NA}.
}
\concept{data_prep_constructors}
