% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{HandleMissingValues}
\alias{HandleMissingValues}
\title{Handle Missing Values in Series Variable}
\usage{
HandleMissingValues(na_remove = FALSE, na_replace = NULL)
}
\arguments{
\item{na_remove}{boolean value, if \code{TRUE} remove \code{NA}s in series variable}

\item{na_replace}{numeric value to replace \code{NA}s in series variable, if NULL
no replacement is performed}
}
\value{
a function to remove or replace missing values.
\subsection{Returned Function}{
\itemize{
\item Arguments
\itemize{
\item \code{data} data frame containing time series data
\item \code{series_variable} column name of series variable in \code{data}, default is "deaths"
}
\item Return - all fields in \code{data} with either \code{NA} records removed or replaced
}
}
}
\description{
Construct a function that takes a data frame and returns another data
frame with \code{NA} values either removed or replaced.
}
\concept{data_prep_constructors}
