% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{HandleZeroValues}
\alias{HandleZeroValues}
\title{Handle Zero Values in Series Variable}
\usage{
HandleZeroValues(zero_remove = FALSE, zero_replace = NULL)
}
\arguments{
\item{zero_remove}{boolean value, if `TRUE` remove zeroes in series variable}

\item{zero_replace}{numeric value to replace zeroes in series variable, if NULL
no replacement is performed}
}
\value{
a function to remove or replace zero values.

## Returned Function

- Arguments
   * `data` data frame containing time series data
   * `series_variable` column name of series variable in `data`, default is "deaths"
- Return - all fields in `data` with either zero records removed or replaced
}
\description{
Remove or replace series variable values that are zero.
}
\concept{data_prep_constructors}
