% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{HandleZeroValues}
\alias{HandleZeroValues}
\title{Handle Zero Values in Series Variable}
\usage{
HandleZeroValues(zero_remove = FALSE, zero_replace = NULL)
}
\arguments{
\item{zero_remove}{boolean value, if \code{TRUE} remove zeroes in series variable}

\item{zero_replace}{numeric value to replace zeroes in series variable, if NULL
no replacement is performed}
}
\value{
A function like \code{\link{handle_zero_values_default}} to remove
or replace zero values.
}
\description{
Construct a function that takes a data frame and returns another data
frame with \code{0} values either removed or replaced.
}
\concept{data_prep_constructors}
