% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_get_metadata}
\alias{iidda_get_metadata}
\title{Get IIDDA metadata}
\usage{
iidda_get_metadata(data, time_variable = NULL, descriptor_variable = NULL)
}
\arguments{
\item{data}{data frame containing time series data}

\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{descriptor_variable}{column name of the descriptor variable in \code{data}, default is "cause" for
mortality data sets.}
}
\value{
a list in order containing minimum time period, maximum time period and cause name.
}
\description{
Get starting time period, ending time period and mortality cause name from the
data set for use in axis and main plot titles.
}
