% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_bar}
\alias{iidda_plot_bar}
\title{Plot Bar Graph}
\usage{
iidda_plot_bar(
  plot_object,
  data = NULL,
  series_variable = "deaths",
  time_unit = "week"
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing data prepped for bar plotting, typically output from `iidda_prep_bar()`.
If `NULL` data is inherited from `plot_object`}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_unit}{time unit to display bar graphs on the x-axis. Defaults to "week" or one of iidda.analysis:::time_units that starts
with "month". Should generalize at some point to be able to take any time_unit argument.}
}
\value{
a ggplot2 plot object containing a bar graphs of time series data
}
\description{
Add a bar plot to an exiting ggplot plot object. Graphical choices were made to closely reflect
plots generated with `LBoM::monthly_bar_graph` and `LBoM::weekly_bar_graph`.
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_ma}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
