% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_plot_ma}
\alias{iidda_plot_ma}
\title{Plot Moving Average Time Series}
\usage{
iidda_plot_ma(
  plot_object,
  data = NULL,
  series_variable = "deaths",
  time_variable = "period_end_date"
)
}
\arguments{
\item{plot_object}{a `ggplot2` plot object}

\item{data}{data frame containing moving average time series data, typically output from `iidda_prep_ma()`.
If `NULL` data is inherited from `plot_object`}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}
}
\value{
a ggplot2 plot object containing a moving average time series
}
\description{
Add a moving average time series line to an exiting ggplot plot object. Graphical choices were made
to closely reflect plots generated with `LBoM::plot.LBoM`.
}
\seealso{
Other plotting_functions: 
\code{\link{iidda_plot_bar}()},
\code{\link{iidda_plot_box}()},
\code{\link{iidda_plot_heatmap}()},
\code{\link{iidda_plot_highlight}()},
\code{\link{iidda_plot_rohani_heatmap}()},
\code{\link{iidda_plot_scales}()},
\code{\link{iidda_plot_series}()},
\code{\link{iidda_plot_settings}()},
\code{\link{iidda_plot_wavelet}()}
}
\concept{plotting_functions}
