% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{iidda_prep_box}
\alias{iidda_prep_box}
\title{Prep Data for Box plot}
\usage{
iidda_prep_box(
  data,
  series_variable = "deaths",
  time_variable = "period_end_date",
  time_unit = "week",
  handle_missing_values = HandleMissingValues(na_remove = FALSE, na_replace = NULL),
  handle_zero_values = HandleZeroValues(zero_remove = FALSE, zero_replace = NULL)
)
}
\arguments{
\item{data}{data frame containing time series data}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}

\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{time_unit}{time unit to create field from `time_variable`. Must be one of iidda.analysis:::time_units, defaults to "week".}

\item{handle_missing_values}{function to handle missing values, defaults to HandleMissingValues}

\item{handle_zero_values}{function to handle zero values, defaults to HandleZeroValues}
}
\value{
all fields in`data` with records prepped for plotting box plots. The name
of the new `time_unit` field will be named from lubridate_funcs.
}
\description{
Prep data for plotting box plots. Prep steps were taken from `LBoM::monthly_box_plot`
and they include handling missing values and creating additional time unit fields.
}
\seealso{
Other prep_data_for_plotting: 
\code{\link{iidda_prep_bar}()},
\code{\link{iidda_prep_ma}()},
\code{\link{iidda_prep_periodogram}()},
\code{\link{iidda_prep_rohani}()},
\code{\link{iidda_prep_seasonal_heatmap}()},
\code{\link{iidda_prep_series}()},
\code{\link{iidda_prep_wavelet}()}
}
\concept{prep_data_for_plotting}
