% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{PeriodAggregator}
\alias{PeriodAggregator}
\title{Period Aggregator}
\usage{
PeriodAggregator(rate_variable, norm_exponent = 5)
}
\arguments{
\item{rate_variable}{Name of variable to be used to store the normalized
count variable.}

\item{norm_exponent}{Exponent to use in normalization. The default is \code{5},
which means \verb{per 100,000}.}

\item{time_variable}{Name of the variable to characterize the temporal
location of the time period.}

\item{period_width_variable}{Name of variable to characterize the width
of the time period.}

\item{count_variable}{Name of variable to characterize the count variable
being normalized.}

\item{norm_variable}{Name of variable to be used to normalize the count
variable.}
}
\description{
Create function that aggregates information over time periods, normalizes
a count variable, and creates new fields to summarize this information.
}
