% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{PeriodAggregator}
\alias{PeriodAggregator}
\title{Period Aggregator}
\usage{
PeriodAggregator(
  time_variable = "period_mid_time",
  period_width_variable = "num_days",
  count_variable = "cases_this_period",
  norm_variable = "population_reporting",
  rate_variable = "daily_rate",
  norm_exponent = 5
)
}
\arguments{
\item{time_variable}{Name of the variable to characterize the temporal
location of the time period.}

\item{period_width_variable}{Name of variable to characterize the width
of the time period.}

\item{count_variable}{Name of variable to characterize the count variable
being normalized.}

\item{norm_variable}{Name of variable to be used to normalize the count
variable.}

\item{rate_variable}{Name of variable to be used to store the normalized
count variable.}

\item{norm_exponent}{Exponent to use in normalization. The default is `5`,
which means `per 100,000`.}
}
\description{
Create function that aggregates information over time periods, normalizes
a count variable, and creates new fields to summarize this information.
}
