% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{SeriesHarmonizer}
\alias{SeriesHarmonizer}
\title{Wavelet Series Harmonizer}
\usage{
SeriesHarmonizer(time_variable = "period_end_date", series_variable = "deaths")
}
\arguments{
\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{series_variable}{column name of series variable in `data`, default is "deaths"}
}
\value{
function to harmonize disease/cause names

## Returned Function

- Arguments
   * `data` data frame containing time series data
- Return - all fields in `data` with summarized series variable for unique time variable
}
\description{
Harmonizes the series variable in `data` so there is one data value for each time
unit in time variable (to account for different variations in disease/cause name)
}
\concept{data_prep_constructors}
