% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{SeriesHarmonizer}
\alias{SeriesHarmonizer}
\title{Wavelet Series Harmonizer}
\usage{
SeriesHarmonizer(sum_fn = base::sum)
}
\arguments{
\item{time_variable}{column name of time variable in \code{data}, default is "period_end_date"}

\item{series_variable}{column name of series variable in \code{data}, default is "deaths"}
}
\value{
function to harmonize disease/cause names
\subsection{Returned Function}{
\itemize{
\item Arguments
\itemize{
\item \code{data} data frame containing time series data
}
\item Return - all fields in \code{data} with summarized series variable for unique time variable
}
}
}
\description{
Harmonizes the series variable in \code{data} so there is one data value for each time
unit in time variable (to account for different variations in disease/cause name)
}
\concept{data_prep_constructors}
