% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{TrimSeries}
\alias{TrimSeries}
\title{Trim Time Series}
\usage{
TrimSeries(zero_lead = FALSE, zero_trail = FALSE)
}
\arguments{
\item{zero_lead}{boolean value, if \code{TRUE} remove leading zeroes in \code{data}}

\item{zero_trail}{boolean value, if \code{TRUE} remove trailing zeroes in \code{data}}
}
\value{
A function like \code{\link{trim_series_default}} to remove to
remove leading and/or trailing zeroes.
}
\description{
Remove leading or trailing zeros in a time series data set.
}
\concept{data_prep_constructors}
