% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{WaveletJoiner}
\alias{WaveletJoiner}
\title{Wavelet Joiner}
\usage{
WaveletJoiner(
  time_variable = "period_end_date",
  series_suffix = "_series",
  trend_suffix = "_trend",
  keep_series_dates = TRUE
)
}
\arguments{
\item{time_variable}{column name of time variable in `data`, default is "period_end_date"}

\item{series_suffix}{suffix to be appended to series data fields}

\item{trend_suffix}{suffix to be appended to trend data fields}

\item{keep_series_dates}{boolean flag to indicate if the dates in `series_data` should
be kept and data from `trend_data` is left joined, if `FALSE` dates from `trend_data`
are left joined instead}
}
\value{
function to join data and trend data sets

## Returned Function

- Arguments
   * `series_data` data frame containing time series data
   * `trend_data` data frame containing trend data
- Return - joined data set by `time_variable` with updated field names
}
\description{
Joins series data and trend datasets and keeps all time units in one of the
datasets.
}
\seealso{
Other data_prep_constructors: 
\code{\link{ComputeMovingAverage}()},
\code{\link{HandleMissingValues}()},
\code{\link{HandleZeroValues}()},
\code{\link{SeriesHarmonizer}()},
\code{\link{TrimSeries}()},
\code{\link{WaveletDeheaper}()},
\code{\link{WaveletInterpolator}()},
\code{\link{WaveletNormalizer}()},
\code{\link{WaveletTransformer}()},
\code{\link{data_prep_constructors}}
}
\concept{data_prep_constructors}
