% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{WaveletTransformer}
\alias{WaveletTransformer}
\title{Wavelet Transformer}
\usage{
WaveletTransformer(
  time_variable = "period_end_date",
  wavelet_variable = "detrend_norm",
  ...
)
}
\arguments{
\item{time_variable}{name of the time variable field in `data`}

\item{wavelet_variable}{name of the field in `data` to be wavelet transformed}

\item{...}{Arguments passed on to \code{\link{analyze.wavelet}}.}
}
\value{
function that computes wavelet transform

## Returned Function

- Arguments
   * `data` data frame containing time series data
- Return - the wavelet transform object from EMD::analyze.wavelet applied to
`wavelet_variable` in `data`
}
\description{
Compute the wavelet transform.
}
\concept{data_prep_constructors}
