% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{WaveletTransformer}
\alias{WaveletTransformer}
\title{Wavelet Transformer}
\usage{
WaveletTransformer(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link{analyze.wavelet}}.}

\item{time_variable}{name of the time variable field in \code{data}}

\item{wavelet_variable}{name of the field in \code{data} to be wavelet transformed}
}
\value{
function that computes wavelet transform
\subsection{Returned Function}{
\itemize{
\item Arguments
\itemize{
\item \code{data} data frame containing time series data
}
\item Return - the wavelet transform object from EMD::analyze.wavelet applied to
\code{wavelet_variable} in \code{data}
}
}
}
\description{
Compute the wavelet transform.
}
\concept{data_prep_constructors}
